Need to provide different Future implementations for submit methods of the new ClientExecutionService (#496)
* Added the IFuture interface and added the same implementation of ClientInvocationFuture as in Java.

* Fix for notifyException. Failing to complete the future sometimes causes the shutdown to get stuck the invocation waiting thread hangs up.

* Introduced exception wrapping and rethrowing to throw the same exceptions as the Java client.

* Fix on connection write after close.

* PnCounterFunctionalityTest::testParallelism test fix. The test did not start the threads as expected.

* Simplified ClientInvocation::isRetrySafeException check by using the newly added isRetryable method to the IException.

* Documentation corrections: removed java related things.

+++ hazelcast/src/hazelcast/client/proxy/IQueueImpl.cpp
@@ -108,7 +108,7 @@ namespace hazelcast {
 
             std::vector<serialization::pimpl::Data> IQueueImpl::drainToData(size_t maxElements) {
                 std::auto_ptr<protocol::ClientMessage> request =
-                        protocol::codec::QueueDrainToMaxSizeCodec::encodeRequest(getName(), maxElements);
+                        protocol::codec::QueueDrainToMaxSizeCodec::encodeRequest(getName(), (int32_t) maxElements);
 
                 return invokeAndGetResult<std::vector<serialization::pimpl::Data>, protocol::codec::QueueDrainToMaxSizeCodec::ResponseParameters>(
                         request, partitionId);
